marimo/pyproject.toml
[project]
name = "pyspark-in-marimo"
version = "0.1.0"
description = "Pyspark examples using Marimo."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "delta-spark>=4.0.0",
    "marimo>=0.18.0",
    "numpy>=2.3.5",
    "pyarrow>=22.0.0",
    "pyspark>=4.0.1",
    "ruff>=0.14.6",
]
